requirements.txt
Django==3.0.8
numpy
channels
editdistance
indic_transliteration
jsonfield
djangorestframework
torch
torchtext
configargparse
nltk
subword-nmt
